Update patch set 1
Patch Set 1:

> Curious.  So the jjb-init-project thing generates both the cfg and
 > yaml, but the cfg can then be changed, which eventually causes
 > generated changes in the yaml?

That's right. Initially we use jjb-init-project to generate the cfg file. Once that's bootstrapped. The builder-merge job always runs jjb-init-project passing the -c or --config option to generate the yaml file.

So we expect projects to only ever modify the cfg file and let Jenkins manage the yaml file as it's auto-generated from cfg.

Patch-set: 1