Update patch set 1
Patch Set 1:

Good to know.  This is out of scope of this review, of course, but I've always felt it important to make sure it's clear what the "protocol" is for usage of generated files.  The obvious step is to add the "don't modify, generated file" comment to the yaml file, which is already there, but it might be helpful to add a comment to the generated cfg saying "generated by jjb-init-project, but can be modified." or something to that effect.

Patch-set: 1