Update patch set 1
Patch Set 1: Code-Review-1

We have a global template [0] you can use now so you don't need to create your own template.

You just need to pass in a toxdir parameter in your project definition if you're tox.ini is not in the root directory of the repo.

[0] https://github.com/opendaylight/releng-builder/blob/master/jjb/releng-templates.yaml

Patch-set: 1
Label: Code-Review=-1