Update patch set 1
Patch Set 1:

Can we name this netvirt-verify-custom so that we can reuse it for other topic branches if needed?

I am also still confused on why the whole template had to be pasted in?

In [1] we removed the ovsdb-verify-custom job since it it no longer needed in the ovsdb project. But in the yaml file they simply add the project info and nothing for the template.

[1] https://git.opendaylight.org/gerrit/#/c/37681/1/jjb/ovsdb/ovsdb-verify-custom.yaml

Patch-set: 1