Update patch set 1
Patch Set 1: Code-Review-1

> This is needed to pass stable/lithium distribution verify job.

This is hiding a stable/lithium distribution Bug.

Is there a Bugzilla item describing the bug and is it set to WONTFIX so that we need to accept the presence of its symptoms?
If yes, add a pointer to code comments.

Patch-set: 1
Reviewer: Gerrit User 2046 <2046@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1