Split cluster node startup and cluster node configuration
The process of installation of ODL to a cluster node takes
way too long, leading to the state when the first node being
installed finishes booting before the second and third nodes
even get around to starting the boot. This is not a supported
cluster startup scenario and thus leads to numerous spurious
test failures which may appear and disappear depending on the
network load on the Nexus server. Namely when the data
transfer rate drops to 100 kb/s before node 2 installation and
stays there until node 2 is started, then node 1 finishes
booting before node 2 and node 3 even get around to download
their copy of ODL, leading to the spurious failures due to the
unsupported scenario.

This change splits the setup script to 2 scripts, the first one
deploys ODL and makes all the configurations and the second one
just runs bin/start. Then the first part of the script is run
on every node and only after all of them are set up, the second
part is run on every node so they start booting at roughly the
same time.

Change-Id: If8b32309a996451fbc4e1a021fdebfe635e0f916
Signed-off-by: Jozef Behran <jbehran@cisco.com>