Update patch set 2
Patch Set 2:

> has completely broken Netconf

Well, to be exact the Netconf in that node works somehow (is able to configure the device in the clustered setup) but looks broken from the point of view of the Robot trying to talk with the node (the Netconf device nor its data is not visible there at all). So it actually isn't completely broken (but looks so).

Patch-set: 2