Update patch set 2
Patch Set 2: -Code-Review

Testing this is proving to be difficult without getting this patch merged. Let's start with a setting of 4 to see how it behaves and watch it for awhile. If we were to increase the cores on the server I would bump the number of cores to 2 so a setting of 4 would work well for 2 cores anyway.

Patch-set: 2