Fix add a New Project Quick Start section
Many projects are confused when they join and are not sure what they
need to do when we refer them to this guide. Let's create a section that
gets them what they need to get the basic jobs going and they can read
the rest of the guide when / if they need to contribute to new JJB
items.

Change-Id: I9d13ce84c7307fa4c575e04b5086e838792d405c
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>