Update patch set 1
Patch Set 1:

I'm wondering how different this .yaml is compared to the already
existing ovsdb-full-integration.yaml?  I see that the ovs versions
only have 2.4 here, but uses 2.3.2 and 2.4 in the full-integration.

I'm asking because if they are very similar in every other way,
could we just add the "runit" gerrit trigger to that 
full-integration.yaml instead of creating this whole new job file?

Patch-set: 1