Update patch set 1
Patch Set 1:

> currently the full-integration job runs on two versions of OVS (2.3
 > and 2.4). Is it possible to use the same full integration job, but
 > change the triggered one to run only on default OVS?

Sharad,

I think the extra time for the full-integration job is because the
timer starts when the full-integration matrix job starts.  But that
job is triggering the two other (2.3 and 2.4) IT jobs.  If the
resources for those jobs aren't ready, then full-integration would
be stuck waiting until those resources are ready.  That's my best
guess.

Patch-set: 1
Label: Code-Review=0