Update patch set 1
Patch Set 1: Code-Review-1

Giving -1 as this is just for verification of 40030 and 39536.

But Sandbox runs were a (relative) success.
The Merge job [0] failed (as it should) only when attempting Nexus upload.
The verify job [1] failed, because the triggered deploy job [2] got "Connection refused.", which I think is a issue of Sandbox configuration and not fault of this Change.

[0] https://jenkins.opendaylight.org/sandbox/job/integration-distribution-merge-boron/1/console
[1] https://jenkins.opendaylight.org/sandbox/job/integration-distribution-verify-boron/11/console
[2] https://jenkins.opendaylight.org/sandbox/job/integration-distribution-deploy-boron/10/console

Patch-set: 1