Update patch set 1
Patch Set 1:

is this right?

I tried to push this job to sandbox, but it wouldn't go with some
error around the gerrit keyword config section which is not part of
this change and we know works in releng already.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0