Update patch set 1
Patch Set 1:

> the settings-type changes are more important

I am confused. Usually the cfp lines only appeared in the Sandbox-related drafts, up until 40042 went in.

But if those are important, why are files in jjb-templates (and jjb/ynagtools/yangtools-verify.yaml) still not updated in master?

This Change is intended to be merged before the Sandbox related ones. I will edit in the lines into 38999 when this is merged.

Patch-set: 1