Update patch set 21
Patch Set 21:

one thing that stands out to me, is if something "networky" is
broken in our infra, this new script is going to spin for 5m per
VM.  In the case where we have 6+ VMs we'll have jobs stuck in
this polling loop for 30m or longer.  

would it work if you switched the loops?  So:

for i in 1..30:
  for vm in $VMLIST:
    do my ssh things

I think that would cap our total wait time on ~5m no matter how many
vms are in the job.

Patch-set: 21
Label: Verified=0