Revert 8G VM changes
It seems the amount of jobs that require 8G VMs is pretty significant
making the majority of VMs that spin up needing to be 8G anyway so we
don't really get much benefit of spinning up the smaller VM types.

The server side has been reverted to running 8G VMs by default for the
dynamic_verify and dynamic_merge labels.

Change-Id: I8c049f4ace9dbe824b1eb06d7b12fe30b8d4fb45
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>