Update patch set 27
Patch Set 27:

> > one thing that stands out to me, is if something "networky" is
 > > broken in our infra, this new script is going to spin for 5m per
 > > VM.  In the case where we have 6+ VMs we'll have jobs stuck in
 > > this polling loop for 30m or longer.
 > >
 > > would it work if you switched the loops?  So:
 > >
 > > for i in 1..30:
 > > for vm in $VMLIST:
 > > do my ssh things
 > >
 > > I think that would cap our total wait time on ~5m no matter how
 > > many
 > > vms are in the job.
 > 
 > I think if we do that we'll need to add some variables to break out
 > of the first loop. Let me think about this...

I think the latest patch is still using the original strategy where it's
looping 30 times, one VM at a time, right?  If so, I guess my initial
comment is still valid.  It's only a problem when we are in a problem
state (bad networking so ssh-copy-id can't work).  In that case, we
are looping 300s per VM all the way until the last VM (maybe 8 or more)
is deemed a failure.  Is there no easy way to break from the 1..30 outer
loop if all VMs are successful?

Patch-set: 27