Update patch set 1
Patch Set 1:

> > [1] is the sandbox job. The added script code to add the bridges
 > > looks to have worked, but then something went wrong with the
 > > disable of firewalld and iptables so the job punted.
 > >
 > > [1] https://jenkins.opendaylight.org/sandbox/job/netvirt-vpnservice-csit-1node-openstack-mitaka-openstack-boron/3/console
 > 
 > yeah, we have to comment out those lines that muck with firewalld. 
 > it's needed in releng, but not sandbox.  once we are migrated, we
 > can
 > get it in sync upstream.

I did comment out the firewalld stuff and ran the job again [0] and it made it to the point where it tried to run robot.  But, we don't have a robot testplan file in place for this new job name.  If we get a gerrit for that, we can run the sandbox job using that gerrit as the
patchrefspec.

[0] https://jenkins.opendaylight.org/sandbox/job/netvirt-vpnservice-csit-1node-openstack-mitaka-openstack-boron/4/console

Patch-set: 1
Label: Code-Review=0