Update patch set 3
Patch Set 3:

the networking-odl part that devstack calls during stacking checks explicitly for netvirt:1. It starts ODL, then looks for netvirt:1, then it connects the ovsdb node and then looks for bridges.

The reason was it took 90s or so for ODL to come up with everything needed. restconf and openflowplugin would come up before netvirt was up. The switches would connect but netvirt wasn't up so it never saw the switches and the pipeline would not be installed.

Patch-set: 3