Update patch set 2
Patch Set 2:

(1 comment)

Hi Vratko, after talking to Thanh in the hackfest, we agreed the best place for elasticsearch installation is vagrant file. If things have not changed jenkins-script is the normal first step before putting something in vagrant.

For your second question, the SW will be installed as root but I believe the user jenkins has sudo access in the controller VM so in the scriptplan we have to write to start the database we will use sudo to start the elasticsearch.

Patch-set: 2