Update patch set 5
Patch Set 5: Code-Review-1

Venkat, why don't you remove the netvirt-vpnservice jobs from this patch for now and we'll merge it.  We'll make a new patch for those jobs and get those merged after we make sure it's working like we think (with the create_tunnels suite) in the sandbox.

Patch-set: 5
Label: Code-Review=-1