Convert dynamic minions to single use
The private cloud minions come online significantly faster than
previously. We also don't get charged anything for creating a lot of
them, so it makes sense to make them single use so we can avoid a lot of
the re-use issues we seem to be seeing lately.

This change modifies the dynamic minions that releng itself uses as well
as the base templates that most projects are using. It does not affect
all the other jobs out there that are not part of the standard template
sets such as robot or custom jobs.

Change-Id: I42237f92e5288c36c34f4748eec6c948ca9553a3
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>