Update patch set 3
Patch Set 3:

> Sorry, but still no success on making the sandbox get the new
 > templates patch. Many times before I've used the PATCHREFSPEC
 > variable to specify a patch belonging to the repo being tested
 > (e.g. sfc or integration/test) but I have found no clue on how to
 > do the same for releng/builder repo patch. I keep on trying :)

what you want to do, is take your local builder repo with these changes and use
that to push your job to the sandbox.  That will create the sandbox job exactly
as this patch describes.

Then, if you want your sandbox CSIT job to use some Int/Test patch which
is not yet merged, that's what PATCHREFSPEC is for.

let me know if it's still not clear.  I know it can be confusing.

Patch-set: 3
Label: Verified=0