Keep build artifacts for 14 days only
The deleting artifacts after 5 max is causing confusion to folks. Lets
reconfigure and simplify the log rotation to keeping max 14 days and
removing the extra configurations.

In the future as we continue to migrate jobs to logs.opendaylight.org
that system has a 6 month log retention so we can reduce Jenkins
further as we proceed.

Change-Id: I98c1c0e76cf4e93244b82415c3091dc11b9f22d8
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>