Update patch set 4
Patch Set 4:

Thanks, finally the docker instance is being deployed (you can see it in [0]). However I have other pending commit to test if this is fully working. I would like to make sure that the docker instance is up&running and accessible.
Thanks for the help! 

[0] https://jenkins.opendaylight.org/sandbox/job/sfc-csit-3node-docker-full-deploy-all-boron/
[1] https://git.opendaylight.org/gerrit/#/c/40600/

Patch-set: 4
Label: Verified=0