Update patch set 2
Patch Set 2:

Could we please do this as two separate changes instead of force overriding Jenkins here?

Change 1) remove the jobs that Jenkins would would failing on
Change 2) remove the files that are causing Jenkins to fail

I don't want to have to remove a vote just to merge and I'm thinking ahead to how folks would have to deal with stuff if Jenkins was the only way to do the merges.

Patch-set: 2
Label: Verified=0