Update patch set 1
Patch Set 1:

Hi,
 I had to create this commit to be able to continue working on [0] and [1]. Sudo is failing in docker minions because of lack of tty/askpass. 

My impresion is that it is not possible to test this modification in a sandbox without merging it first (unlike other modifications in yaml templates, jenkins-jobs does not get it when creating the job in the sandbox). 

If there is a way to do it please let me know and accept my apologies.
 
Thanks for your help.

[0] https://git.opendaylight.org/gerrit/40582
[1] https://git.opendaylight.org/gerrit/40600

Patch-set: 1
Label: Verified=0