Update patch set 7
Patch Set 7:

Thanks Jamo, but I'm not sure about that. I just introduced intentionally a bunch of errors in jenkins-scripts files and then push a new job that makes use of them and it worked fine (and it should have failed). It seems that jenkins-scripts are not being pushed to the sandbox. I'm double checking this to make sure I didn't do anything wrong.

Patch-set: 7