Update patch set 4
Patch Set 4:

Sorry, but I'm a bit confused now: I intentionally added a bunch of errors in every script of 'jenkins-script' spin-ups and then push a new sandbox job hoping that it will fail but it didn't. It seems that the sandbox job does not get the jenkins-scripts modifications (unlike job templates and yaml files that are correctly included and executed).

I created this commit to have this modification merged separately but if there is not verification process the commit will not be merged (and I don't see a "+1 verify" button, but maybe commiters do). 

Yesterday, when Thanh Ha gave +2 the commit was still expecting a "+1 verify" that never appeared.

Maybe commiters can give "+1 verify" (aside from +2 CR) ?

Again sorry if this is due to my misunderstanding, still double checking everything just in case. :)

Patch-set: 4