Update patch set 1
Patch Set 1:

The reason is to make it more obvious which dependencies are being pulled in, in attempt to reduce version skew. But I am not sure yet on how to write it into commit message in a way that would prompt every project with MVN_GOALS in their .cfg file to give +1.

Patch-set: 1