Update patch set 1
Patch Set 1: Code-Review-1

cfg files don't do anything anymore as of last night and will be removed in the near future once we confirm everything is well with the world.

https://git.opendaylight.org/gerrit/41681

Basically all job templates are now regular jjb templates so please modify the project.yaml file directly. (or one of the templates in releng-templates-java.yaml)

Patch-set: 1