Update patch set 1
Patch Set 1:

> please modify the project.yaml file directly

Oh, I see now. By the way, is there a reason why the following text is in {project}.yaml and not in releng-templates-java.yaml?

-Dmaven.repo.local=/tmp/r -Dorg.ops4j.pax.url.mvn.localRepository=/tmp/r

If there is a plan to move that soon, could dependency:tree be also added there, or would it be better to keep it configurable on {project}.yaml level?

Patch-set: 1