Update patch set 2
Patch Set 2:

The -DoutputFile might conflict with another task I want to do. We want to move all jobs to freestyle jobs away from Maven Project jobs and part of that will will involve redirecting module output to a file. So if dependency:tree is outputing to a file too we will now have 2 things that want to output to the file.

I think -DoutputFile creates a log containing all maven output if I recall right? so maybe this can be more generically renamed as $MODULE.log or something and logged to somewhere in the $WORKSPACE/archives directory. I want to create a full log of each individual module like what we get from the Maven Project and generate an index page that has convenient links to each module with the output of the Maven reactor summary at the end of the build.

This will give similar convenience to what the Maven Project job type gives us without needing to use it and would speed up our builds. I plan to create a mailing list thread soon to discuss this.

Patch-set: 2