Update patch set 1
Patch Set 1: Code-Review+2

> JJB translates this

Oh, I had wrong idea of what the "raw" in include-raw stands for. After checking documentation, I assume include-raw-escape is what I was confusing it for (hard to tell, as the documenation does not show the processed data).

Anyway, I checked a real job on Sandbox and you are right.

[0] http://docs.openstack.org/infra/jenkins-job-builder/definition.html

Patch-set: 1
Label: Code-Review=+2