Update patch set 1
Patch Set 1: Code-Review+1

Sam, Anil,

this patch was rechecked to show a comparison of the new freestyle type
vs matrix:

https://git.opendaylight.org/gerrit/#/c/42059/

the freestyle (because this was beryllium) kicked two jobs (one for
java7 and one for java8).

https://jenkins.opendaylight.org/releng/job/ovsdb-verify-beryllium-mvn33-openjdk7/9/
https://jenkins.opendaylight.org/releng/job/ovsdb-verify-beryllium-mvn33-openjdk8/2/

the old matrix job runs both as well, but under the umbrella of a 
single job:

https://jenkins.opendaylight.org/releng/job/ovsdb-verify-beryllium/2361/

The freestyle jobs run a few mins faster.

We can wait for a +1 from either of you guys before merging.  Thanh
wants to get this working for all projects and ovsdb is the guinea pig.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1