Update patch set 1
Patch Set 1:

there is probably going to be a window when this change is there
for the job, but the distribution patches are not merged to make
this feature available in the distro this job takes.

those other patches are merged, but their merge job failed.

typical window we have to deal with in these situations.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0