Convert Vagrant to Packer
Changes in the Rackspace Public cloud that came about because of the
private cloud creation and connection have made using Vagrant inside the
public cloud impossible now. This converts all of the in use Vagrant
definitions to Packer and allows us to build images at the same time in
both clouds so that they are identical as we can possibly make them.

Change-Id: I410ad83eaabddfe50447dbeefe7eb2d79660a06c
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>