Migrate yangtools matrix-verify jobs to freestyle-verify jobs
This patch will:

    * Delete the existing yangtools verify jobs
    * Create new template {project-name}-verify-{stream}-{maven}-{jdk} jobs
    * Switch yangtools back to use a releng managed verify template

In the case of yangtools we are switching from the Maven project job
type to freestyle project type. It should also speed up builds a little
at the end as the freestyle jobs do not run the additional
fingerprinting.

This work is related to our infra changes mentioned on the mailing list
(see below for link). We'd like to migrate jobs away from using the
matrix type jobs and convert them to freestyle jobs as the matrix system
occassionally causes us build issues especially with larger queue sizes.

See:
https://lists.opendaylight.org/pipermail/release/2016-July/007288.html

Change-Id: I6a2fb2d7745a089f29106defd6e4fe4d59584c9c
Co-Authored-By: Thanh Ha <thanh.ha@linuxfoundation.org>
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>