Split OVSDB IT into separate mvn targets
Previously there was only one maven target for OVSDB IT. This
would generate one large file, with the logs from all of the
tests.

This patch creates 4 separate maven targets for OVSDB IT:
  - build
  - library/it
  - southbound-it
  - hwvtepsouthbound-it

Each target will have  it's own log file. This should make
it a little easier to debug each particular IT test target.

Also, this patch adds a new file to config docker for hwvtep
emulation. This config will only occur for OVSDB IT tests
that need it, currently library/it and hwvtepsouthbound-it.

Change-Id: I49a85bf6055e3233747e822038a78b2261e8e24c
Signed-off-by: Victor Pickard <vpickard@redhat.com>