DO NOT MERGE
This patch splits OVSDB into the following 4 targets:

1. Build OVSDB

For each OVS version:
2. Run library IT
3. Run southbound IT
4. Run hwvtepsouthbound IT

With this approach, I'm concerned that the if the job
fails on #2 above, for OVS version 2.3.3, then the
job will terminate, and not run for other OVS versions.

I don't think that is what we want. I think we
would want to run #2-#4 for each OVS version independently,
which would mean possibly going back to matrix job.

Change-Id: Ifff0353f7176a7dec1b7d4f10e2d5ee4a72d7b59
Signed-off-by: Victor Pickard <vpickard@redhat.com>