Update patch set 2
Patch Set 2:

> Did you get a chance to run this in the sandbox? I am curious to
 > see how this looks.

Hi, 
I was able to test it in sandbox you can see the results in [1].
It seems that HTML publisher plugin does not support CSS so the results looks very bad. Need to save tempest_results.html and then you will be able to see how it should look. In order to resolve it, we configure our local Jenkins to support CSS see [2], can we configure it there too ? 

[1]
https://jenkins.opendaylight.org/sandbox/view/All/job/netvirt-csit-1node-openstack-mitaka-openstack-boron/2/robot/report/
[2]
http://stackoverflow.com/questions/35783964/jenkins-html-publisher-plugin-no-css-is-displayed-when-report-is-viewed-in-j

Patch-set: 2