Update patch set 8
Patch Set 8:

Finally this can be checked out in [0].

Please have a look. I know there are changes pending to be updated but since it took me really long to have it working I wonder if it is possible submit it now and continue working. 

Thanks,
Best Regards,
Miguel Ángel.
[0] https://jenkins.opendaylight.org/sandbox/job/sfc-csit-3node-docker-full-deploy-all-boron/3/

Patch-set: 8
Label: Verified=0