Update patch set 2
Patch Set 2:

Hi Rob, can you review and +1 this patch if you're okay with it?

We plan on getting rid of the matrix master as soon as all jobs that use it stop as matrix jobs are causing issues with our CI so we want to make all job types be freestyle.

Patch-set: 2