Update patch set 1
Patch Set 1:

CluterKeywords assume the variable is already present, because our scripts provide it for 3node jobs. ClusterManagement init keyword defaults to 1 if the variable is not present.

I thought that only suites using ClusterKeywords would need this Change, but now I see there are also scrips such as csit/scripts/set_elasticsearch_attribute_long.sh referencing the variable.

Commit message does not help with figuring out what the intent of this Change is, so I assumed the worst. :)

Patch-set: 1