Update patch set 1
Patch Set 1:

> Oh, right. Jobs like [0] only work because they do NOT attempt to
 > clone {project} repository. And even those should be renamed to
 > prevent confusion.
 > 
 > [0] https://jenkins.opendaylight.org/releng/view/netvirt/job/netvirt-vpnservice-csit-1node-openstack-mitaka-openstack-boron/


well, this does work.  there are two netvirt impls.  one comes from
the vpnservice project.  so we have "netvirt" and we have "netvirt-vpnservice".  we differentiate in the jjb {project} name.  same thing
that happened in openflowplugin with -li, but -li was done in the {functionality} part of the jjb.  both work.  probably {functionality} is better, but we are already doing it in {project} from netvirt in other places, so this stays with that format.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0