Update patch set 2
Patch Set 2:

Hi Luis
Counters can be use to indicate if an action in the system passed through an expected flow and to know the number of times an operation occurred. The counters will be written in the log in a configurable interval. Diff counters, the default type of counters, will be written with the difference in value from the last they were written. State counter will be written with the current total value it has since the system start. Counters are per ODL instance and each ODL log will contain only its counters.

It was introduced in the Infra Utils porject, and currently netvirt is the first project using it.
You can see some explenaton here:
https://wiki.opendaylight.org/view/Getting_started_with_Counters

and a lecture about it from the last DDF:
https://www.youtube.com/watch?v=gzf-sSdtOrc

Patch-set: 2
Label: Verified=0