Update patch set 3
Patch Set 3:

> actually this patch is not quite right yet.  we need to get the
 > template mapping correct before doing it.  we don't need a new
 > template though.  maybe we can use the string "legacy" in the
 > functionality variable of the job yaml?

what do you mean by template mapping correct? is there a reference to the template name somewhere in a file not included?

we can use the functionality, but I don't like the asymmetric way of the naming - the new netvirt is netvirt-vpnservice-csit-openstack-integration

maybe do a commit to rename all functionalities to either legacy or vpnservice and use a single template for both runs? feel free to patch this review if I am missing something, not that familiar with csit framework

Patch-set: 3