Do not archive maven sites
Maven sites are published to Nexus under the site url here:
https://nexus.opendaylight.org/content/sites/site/

VTN should not be archiving Maven sites in Jenkins or
logs.opendaylight.org as it affects the stabilityi of Jenkins.

Additionally Javadoc jars are also published to Nexus already do not
archive them in Jenkins or logs.opendaylight.org.

Change-Id: I941e4ea561f6dc14baa607a237e979cfd62e772a
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>