Update patch set 4
Patch Set 4:

I gave this one more thought and I think it's too risky. Some projects have ~1hr build times and we cannot have partial artifacts pushed to Nexus for that long duration. It's possible that a new artifact will break downstream projects so we really need a solution that lets us deploy the new artifacts in one go at the end of the job.

Patch-set: 4