Update patch set 1
Patch Set 1: Code-Review-1

This will not work, some jjb templates do not use the project value but this template will use the project name to download project repository. In general I do not recommend to fill project value with anything that is not real project name although this may be late for some of the existing jobs.

Patch-set: 1
Label: Code-Review=-1