Update patch set 1
Patch Set 1: Code-Review-1

Bundle files are exclusive to autorelease-release jobs. I think the solution should be done in such a way that it only applies to the autorelease release job. The releng-* macros should be designed generically that applies broadly to all projects, we shouldn't add things that are specific to one project in it.

Patch-set: 1
Label: Code-Review=-1