Update patch set 1
Patch Set 1:

(1 comment)

I'm not sure this is necessary. tox without any parameters will run the "envlist" in the tox.ini file. Docs needs to run 2 testenvs "docs" and "docs-linkcheck" so using this we'd be skipping one of them.

I think I prefer to run the default envlist and let projects define what that is in their tox.ini unless you have a reason to not do this?

Patch-set: 1
Label: Verified=0