Introduce verify-tox to replace verify-python over time
The python-verify job has been executing "tox",
which runs tox environments specified in envlist.
At first, this was only "pep8" (hence python),
but now it can be "docs" or combinations with anything else.

Because of that, "python" is no longer a correct description in general.

This Change introduces template with "tox" instead of "python"
in its name, They are aliases to each other so are identical except in
name.

Change-Id: Ibe2121281ba24b4f6a102900721e57ff7a6977be
Signed-off-by: Vratko Polak <vrpolak@cisco.com>