Update patch set 1
Patch Set 1: Code-Review+1

Yes I would be happy to see newton runs already, now that we removed liberty. Newton will be out end of month.

This is also required for starting testing of networking-odl_v2..

Isaku, is there an ETA for cutting stable branch of newton for networking-odl? Currently the proposal in this review is to use devstack stable/newton with networking-odl master, but it would be better if we had a stable/newton for networking-odl already.

Patch-set: 1
Reviewer: Gerrit User 670 <670@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1