Fix integration-test downstream trigger in autorelease
Autorelease-release-{stream} jobs were not triggering downstream
integration-test triggers. This was due to the introduction of
opendaylight-infra-ship logs which moves files for archiving out of the
$WORKSPACE and into the archives directory for shipping to
logs.opendaylight.org. While this is normally fine, Parameterized build
triggers actually are designed to run as a final step in a build rather
than earlier so we need to ensure the input file is still around at the
end of the build.

The solution presented here copies the autorelease variables.prop to a
new variables.jenkins-trigger. One copy is for archiving while the other
copy is for triggering the final build.

Change-Id: Idcf419378fd8f60fab653350be856d60f7ba19aa
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>