Update patch set 2
Patch Set 2:

> (1 comment)

Tested the command in mitaka, does not work.  Actually, we have not used this information for any debugging so far. Also, we have the entire logs now archived. So, It is possible to identify the state of services.

Patch-set: 2
Reviewer: Gerrit User 824 <824@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0