Update patch set 1
Patch Set 1:

Initially I had the same though about test-genius but thought it was confusing without the third identifier. Almost made me think that is was the all case since it didn't have the third id. I am fine with either way though.

What is the difference between the only and the all tests?

Patch-set: 1