Update patch set 1
Patch Set 1:

> Please split to -legacy-cluster and current-cluster or just
 > cluster, legacy will slowly die away..
 > 
 > Tentative +1 after change if you agree so you can merge today

we have the parent patch for the rename/gating jobs to merge first, but as for this
one this creates a gerrit keyword "test-netivrt-cluster" and runs those 3 suites.  I
was thinking that clustering was more core to both impls, but if you think we need
to segregate legacy vs current for cluster too, then I can do:

test-netvirt-legacy-cluster
test-netvirt-current-cluster

I'm fine either way.

Patch-set: 1
Label: Verified=0