Update patch set 1
Patch Set 1:

> We are interested in running jobs both using our git repo (e.g. a
 > different GIT_BASE), but also running jobs using the real ODL git
 > repo (e.g. the default GIT_BASE), for example for third party CI in
 > netvirt.
 > 
 > Currently as I understand it GIT_BASE is only a global parameter in
 > Jenkins, so I cannot control this on a job-by-job basis.
 > 

You can define a different SCM macro for when you want to use a different URL. See:
https://github.com/opendaylight/releng-builder/blob/master/jjb/releng-macros.yaml#L109-L119

Then at the job level you can switch the SCM as necessary.

 > Do you have another suggestion on how to do this?

As mentioned above you should configure a different SCM macro then you can pick and choose which macro you want to use for each job. This way it won't affect us upstream, and you can modify things as you wish downstream. It might make sense to create a HPE specific macros file in your copy of releng/builder so that you can store all the HPE specific tweaks in there.

 > Are you worried only about exposing this variable per-job, or the
 > way I implemented it? (because I can implement the parameters
 > differently)

I'm worried above unnecessarily exposing this URL because in ODL infra we want to make sure that it is easily controlled by Helpdesk so that we can make project wide changes quickly when necessary. If we expose GIT_BASE as a configurable parameter and then projects start using it I'm worried it'll make it more complicated for Helpdesk to manage.

Patch-set: 1