Simplify val-autorelease and distro-check state definition
The validate-autorelease and distribution-check jobs use a
activate/disable mechanism from before when releng/builder was using a
custom scripting layer. This is no longer needed as projects can just
remove the unneeded templates directly in their 'project' definition
jjb so let's simplify the JJB code by removing the complexity here.

Also simplify streams to no longer be a list as that is no longer
necessary with the way we are now defining project definitions.

- templates now have a setting of 'disabled: false' to ensure they are
  always enabled if defined.
- remove validate-autorelease and distro-check templates from projects
  that are not participating in the distribution / autorelease.
- enable validate-autorelease job on projects that are participating in
  the distribution / autorelease but mistakenly had those jobs disabled.

Change-Id: Ia4edb4ad0207ddcf9ca4508886fb3f37488f4788
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>