Update patch set 1
Patch Set 1: Code-Review+1

> Yes, I think these are the subset of tests that are used to
 > validate the release.
 > 
 > The files confuse me also.
 > 
 > Jamo, is it possible to add a comment section in these files to
 > document what they do?

yeah, these are the CSIT jobs that get automatically triggered by autorelease (and
other stuff) and we use those to validate the release.  It's up to each project to 
put their *STABLE* CSIT in these lists.  It's actually part of the release requirements
to put your CSIT here:

https://wiki.opendaylight.org/view/Simultaneous_Release:Boron_Release_Plan#Boron_Project_Integration_and_Test_Requirements

we need to be better about that in netvirt.

as for the comments explaining things, check line 277.  is that what you mean?

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1