Update patch set 9
Patch Set 9:

> (1 comment)
 > 
 > Change to -1 due to Vratko's finding. This means we should
 > problably move the triggers to the build section (not publisher).

Luis, Appreciate your suggestions, presently I don't see an option available to enable block when the build triggers from the publisher section. Having said this, if we move build trigger back to builder section archiving of dist-karaf may not be available for the downstream job, which could potentially cause the build to fail. Not sure if we should consider keeping the integration jobs as is - `maven type` - for the time being unless we have better alternative ?

Patch-set: 9