Update patch set 1
Patch Set 1: Code-Review+1

> > (1 comment)
 > >
 > > The patch looks fine to me.
 > >
 > > One question I had about the job we are removing. Since it is
 > > checking neutron changes as well networking-odl does this changes
 > > our thoughts about eliminating rather than fixing? ?
 > >
 > > This job is currently a resource hog, running every 2-5 minutes
 > on
 > > average and failing https://jenkins.opendaylight.org/releng/view/netvirt/job/netvirt-openstack-gerrit/
 > >
 > > But do we want a job that votes upstream neutron/networking-odl
 > > patches or is current coverage in both CI's adequate? I only ask
 > > this now because I didn't know this was running against neutron
 > > patches as well.

this would be nice, but I don't know if we have the bandwidth to handle
it in ODL infra.  The existing openstack jobs we'd need to run to vote
on networking-odl take 4 VMs and over an hour to run.  If there are are
lot of networking-odl patches we probably can't handle it.


 > 
 > Correction: Maybe the job running 2-5 minutes is not quite accurate
 > but it does seem to run a least once an hour and often more.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1